Initial commit
This is not the very first version - for some odd reason,
it wasn't tracked with Git for a long while. However, this
is quite close to original, except for switch to Go Modules
and addition of `dburl` library to support working with
not just SQLite3 but PostgreSQL as well. Which is not exactly
backwards compatible, because now `janitor.database` requires
extra "sqlite:" prefix.